0.6 -> main

+++ core/kv/KvServer.h
@@ -18,19 +18,25 @@
 namespace nemesis { namespace core { namespace kv {
 
 
-static inline kv::KvHandler * s_kvHandler {nullptr}; // yuck, but easy solution for us_create_timer() 
-
-
-class KvServer
+template<bool HaveSessions>
+class Server
 {
+  struct Handler
+  {
+    KvHandler<HaveSessions> * handler {nullptr};
+  } ;
+
+  static inline Handler s_kvHandler;  // used by the callback handler for us_timer_set(), a C library
+
+
 public:
-  KvServer() : m_run(true)
+  Server() : m_run(true)
   {
     kv::serverStats = nullptr;
   }
 
 
-  ~KvServer()
+  ~Server()
   {
     stop();
   }
@@ -51,8 +57,8 @@ public:
       for(auto sock : m_sockets)
         us_listen_socket_close(0, sock);
 
-      if (s_kvHandler)
-        delete s_kvHandler;
+      if (s_kvHandler.handler)
+        delete s_kvHandler.handler;
 
       if (kv::serverStats)
         delete kv::serverStats;
@@ -66,7 +72,7 @@ public:
     m_wsClients.clear();
     m_sockets.clear();
     m_monitorTimer = nullptr;
-    s_kvHandler = nullptr;
+    s_kvHandler.handler = nullptr;
     kv::serverStats = nullptr;
   }
 
@@ -76,14 +82,18 @@ public:
     if (!init(config.cfg))
       return false;
 
-    if (config.load())
-    {      
-      if (auto [ok, msg] = load(config); !ok)
+    if constexpr(HaveSessions) 
+    {
+      if ((config.load()))
       {
-        PLOGF << msg;
-        return false;
+        if (auto [ok, msg] = load(config); !ok)
+        {
+          PLOGF << msg;
+          return false;
+        }
       }
     }
+    
 
     const unsigned int maxPayload = config.cfg["kv"]["maxPayload"].as<unsigned int>();
     const std::string ip = config.cfg["kv"]["ip"].as_string();
@@ -104,34 +114,46 @@ public:
   
   private:
 
+    // used by uWS timer callback, a C function
+    static void onMonitor (struct us_timer_t *)
+    {
+      if constexpr (HaveSessions)
+        s_kvHandler.handler->monitor();
+    }
+
+
     bool init(const njson& config)
     {
-      if (NemesisConfig::kvSaveEnabled(config))
+      if constexpr (HaveSessions)
       {
-        // test we can write to the kv save path
-        if (std::filesystem::path path {NemesisConfig::kvSavePath(config)}; !std::filesystem::exists(path) || !std::filesystem::is_directory(path))
+        if (NemesisConfig::kvSaveEnabled(config))
         {
-          PLOGF << "kv:session::save::path is not a directory or does not exist";
-          return false;
-        }
-        else
-        {
-          const auto filename = createUuid();
-          std::filesystem::path fullPath{path};
-          fullPath /= filename;
-
-          if (std::ofstream testStream{fullPath}; !testStream.good())
+          // test we can write to the kv save path
+          if (std::filesystem::path path {NemesisConfig::kvSavePath(config)}; !std::filesystem::exists(path) || !std::filesystem::is_directory(path))
           {
-            PLOGF << "Cannot write to kv:session::save::path";
+            PLOGF << "kv:session::save::path is not a directory or does not exist";
             return false;
           }
           else
-            std::filesystem::remove(fullPath);
+          {
+            const auto filename = createUuid();
+            std::filesystem::path fullPath{path};
+            fullPath /= filename;
+
+            if (std::ofstream testStream{fullPath}; !testStream.good())
+            {
+              PLOGF << "Cannot write to kv:session::save::path";
+              return false;
+            }
+            else
+              std::filesystem::remove(fullPath);
+          }
         }
       }
+      
 
       kv::serverStats = new kv::ServerStats;
-      s_kvHandler = new kv::KvHandler {config};
+      s_kvHandler.handler = new kv::KvHandler<HaveSessions> {config}; // TODO
 
       return true;
     }
@@ -146,6 +168,8 @@ public:
 
       auto listen = [this, ip, port, &listening, &startLatch, maxPayload, stats]()
       {
+        char requestBuffer[NEMESIS_KV_MAXPAYLOAD] = {};
+        
         auto wsApp = uWS::App().ws<WsSession>("/*",
         {
           .compression = uWS::DISABLED,
@@ -158,47 +182,36 @@ public:
           {
             m_wsClients.insert(ws);
           },          
-          .message = [this, stats](KvWebSocket * ws, std::string_view message, uWS::OpCode opCode)
-          {   
+          .message = [this, stats, &requestBuffer](KvWebSocket * ws, std::string_view message, uWS::OpCode opCode)
+          { 
             ++serverStats->queryCount;
 
             if (opCode != uWS::OpCode::TEXT)
               ws->send(createErrorResponse(RequestStatus::OpCodeInvalid).to_string(), kv::WsSendOpCode);
             else
             {
-              // TODO this avoids exception on parse error, but not clear how to create json object
-              //      from reader output (if even possible)
-              // jsoncons::json_string_reader reader(message);
-              // std::error_code ec;
-              // reader.read(ec);
-
-              njson request = njson::null();
-
-              try
-              {
-                // TODO look at pmr parse: https://github.com/danielaparker/jsoncons/blob/master/doc/Examples.md#parse-a-json-text-using-a-polymorphic_allocator-since-01710
-                request = njson::parse(message);
-              }
-              catch (...)
-              {
-                // handled below, would have to check request.is_null() anyway
-              }
-                            
+              jsoncons::json_decoder<njson> decoder;
+              jsoncons::json_string_reader reader(message, decoder);
               
-              if (request.is_null())
+              if (reader.read(); !decoder.is_valid())
                 ws->send(createErrorResponse(RequestStatus::JsonInvalid).to_string(), kv::WsSendOpCode);
-              else if (request.empty() || !request.is_object()) //i.e. top level not an object
-                ws->send(createErrorResponse(RequestStatus::CommandSyntax).to_string(), kv::WsSendOpCode);
-              else if (request.size() > 1U)
-                ws->send(createErrorResponse(RequestStatus::CommandMultiple).to_string(), kv::WsSendOpCode);
               else
               {
-                const auto& commandName = request.object_range().cbegin()->key();
+                njson request = decoder.get_result();
 
-                if (!request.at(commandName).is_object())
-                  ws->send(createErrorResponse(commandName+"_RSP", RequestStatus::CommandType).to_string(), kv::WsSendOpCode);
-                else if (const auto status = s_kvHandler->handle(ws, commandName, std::move(request)); status != RequestStatus::Ok)
-                  ws->send(createErrorResponse(commandName+"_RSP", status).to_string(), kv::WsSendOpCode);
+                if (request.empty() || !request.is_object()) //i.e. top level not an object
+                  ws->send(createErrorResponse(RequestStatus::CommandSyntax).to_string(), kv::WsSendOpCode);
+                else if (request.size() > 1U)
+                  ws->send(createErrorResponse(RequestStatus::CommandMultiple).to_string(), kv::WsSendOpCode);
+                else
+                {
+                  const auto& commandName = request.object_range().cbegin()->key();
+
+                  if (!request.at(commandName).is_object())
+                    ws->send(createErrorResponse(commandName+"_RSP", RequestStatus::CommandType).to_string(), kv::WsSendOpCode);
+                  else if (const auto status = s_kvHandler.handler->handle(ws, commandName, request); status != RequestStatus::Ok)
+                    ws->send(createErrorResponse(commandName+"_RSP", status).to_string(), kv::WsSendOpCode);
+                }
               }
             }
           },
@@ -228,23 +241,21 @@ public:
 
         if (!wsApp.constructorFailed())
         {
-          #ifndef NDB_UNIT_TEST
-
-          const auto periodMs = chrono::duration_cast<chrono::milliseconds>(chrono::seconds {5});
-          const auto evtLoop = uWS::Loop::get();
-
-          PLOGD << "Creating monitor timer for " << periodMs.count() << "ms";
-
-          m_monitorTimer = us_create_timer((struct us_loop_t *) evtLoop, 0, 0);
-
-          us_timer_set(m_monitorTimer, [](struct us_timer_t *) mutable
+          if constexpr (HaveSessions)
           {
-            s_kvHandler->monitor();
+            #ifndef NDB_UNIT_TEST
 
-          }, periodMs.count(), periodMs.count());
+            const auto periodMs = chrono::duration_cast<chrono::milliseconds>(chrono::seconds {5});
 
-          #endif
-        
+            PLOGD << "Creating monitor timer for " << periodMs.count() << "ms";
+
+            m_monitorTimer = us_create_timer((struct us_loop_t *) uWS::Loop::get(), 0, 0);
+
+            us_timer_set(m_monitorTimer, Server<HaveSessions>::onMonitor, periodMs.count(), periodMs.count());
+
+            #endif
+          }
+                  
           wsApp.run();
         }
           
@@ -278,6 +289,7 @@ public:
 
 
     std::tuple<bool, std::string> load(const NemesisConfig& config)
+      requires(HaveSessions)
     {
       const fs::path root = config.loadPath / config.loadName;
 
@@ -316,7 +328,7 @@ public:
         return {false, "Cannot load: save was incomplete"};
       else
       {
-        const auto loadResult = s_kvHandler->internalLoad(config.loadName, data);
+        const auto loadResult = s_kvHandler.handler->internalLoad(config.loadName, data);
         const auto success = loadResult.status == RequestStatus::LoadComplete;
 
         PLOGI << "-- Load --";
@@ -345,6 +357,10 @@ public:
     std::atomic_bool m_run;
     us_timer_t * m_monitorTimer{nullptr};
 };
+
+
+using KvServer = Server<false>;
+using KvSessionServer = Server<true>;
 
 
 }
